requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
Pillow==9.5.0
uuid==1.30
python-dotenv==0.19.0
email-validator==1.1.3
gunicorn==20.1.0